requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.18.3
nltk==3.8.1
numpy==2.2.0
openai==0.28.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
playwright==1.49.1
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
selenium==4.27.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
